pyproject.toml
[project]
name = "genai-session9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=3.7.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "openai>=2.7.1",
    "presidio-analyzer>=2.2.360",
    "presidio-anonymizer>=2.2.360",
    "python-dotenv>=1.2.1",
    "ragas>=0.3.8",
]
